Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pillow
pytesseract
opencv-python
sqlalchemy
openai
